Website/requirements.txt
Flask
psycopg2
python-dotenv
gunicorn==23.0.0